+++ types/isSubsetOf.d.ts
@@ -1,0 +1,7 @@
+import type {ImmutableTree} from './types';
+
+export default function isSubsetOf<T>(
+  tree: ImmutableTree<T>,
+  other: ImmutableTree<T>,
+  cmp: (a: T, b: T) => number,
+): boolean;